a lot of bcc64x refactoring

+++ Source/Forms/fSettings.cpp
@@ -51,7 +51,7 @@ void __fastcall TfrmSettings::FormCreate(TObject *Sender)
 //---------------------------------------------------------------------------
 void __fastcall TfrmSettings::btnProfileClick(TObject *Sender)
 {
-    pgcSettings->TabIndex = static_cast<TControl*>(Sender)->Tag;
+    pgcSettings->TabIndex = static_cast<int>(static_cast<TControl*>(Sender)->Tag);
 }
 //---------------------------------------------------------------------------
 void __fastcall TfrmSettings::btnOkClick(TObject *Sender)
